Changes corresponding to u128 changes to Bytes

+++ src/engine/strat_engine/backstore/blockdevmgr.rs
@@ -306,7 +306,7 @@ impl BlockDevMgr {
             current_time
         };
 
-        let data_size = Bytes(metadata.len() as u64);
+        let data_size = Bytes(metadata.len() as u128);
         let candidates = self
             .block_devs
             .iter_mut()
